Some refactors and optimizations in vertex layout workflow, not working yet

+++ Assets/Scripts/Chunk/Chunk.cs
@@ -4,5 +4,5 @@ public static class Chunk
 {
     public const byte X_SIZE = 16;
     public const byte Y_SIZE = 16;
-    public const int TOTAL_SIZE = (X_SIZE * Y_SIZE);
+    public const ushort TOTAL_SIZE = (X_SIZE * Y_SIZE);
 }
